+++ core-ng/src/main/java/core/framework/internal/http/BodyLogParam.java
@@ -12,14 +12,15 @@ import static java.nio.charset.StandardCharsets.UTF_8;
  * @author neo
  */
 public final class BodyLogParam {
-    public static Object param(byte[] body, ContentType contentType) {
+    public static Object of(byte[] body, ContentType contentType) {
         if (contentType != null) {
             if (APPLICATION_JSON.mediaType.equals(contentType.mediaType)) {
                 return new JSONLogParam(body, contentType.charset().orElse(UTF_8));    // make json body filterable
             } else if (contentType.mediaType.contains("text")
-                    || contentType.mediaType.contains("xml")    // for application/xml
-                    || APPLICATION_FORM_URLENCODED.mediaType.equals(contentType.mediaType)) {
+                    || contentType.mediaType.contains("xml")) { // for application/xml
                 return new BytesLogParam(body, contentType.charset().orElse(UTF_8));
+            } else if (APPLICATION_FORM_URLENCODED.mediaType.equals(contentType.mediaType)) {
+                return new FormLogParam(body);
             }
         }
         return "byte[" + body.length + "]";