+++ core-ng/src/main/java/core/framework/internal/web/request/RequestParser.java
@@ -160,7 +160,7 @@ public final class RequestParser {
         var body = exchange.getAttachment(RequestBodyReader.REQUEST_BODY);
         if (body != null) {
             request.body = body.body();
-            logger.debug("[request] body={}", BodyLogParam.param(request.body, request.contentType));
+            logger.debug("[request] body={}", BodyLogParam.of(request.body, request.contentType));
         } else {
             parseForm(request, exchange);
         }